// packages/core-client/src/generated/Notice.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CoreEvent } from "./CoreEvent";
import type { Progress } from "./Progress";

/**
 * The single stream from core to UI.
 */
export type Notice =
  | { type: "event"; event: CoreEvent }
  | { type: "progress"; op: number; progress: Progress };